RBAC tests: extract projectCreated to break platform.v3.server cycle (TRI-8731)
After rebasing rbac-packages on origin/main, the e2e.full harness
regressed with the same `TypeError: Cannot convert undefined or null
to object at allMachines (build/index.js:71862)` that TRI-8731
worked around by switching the testcontainer to NODE_ENV=production.
Recent main commits changed the bundled init order enough that the
production-mode dodge no longer applies — the cycle now triggers in
both NODE_ENV=test and NODE_ENV=production.

Root cause is structural:

  app/services/platform.v3.server  →  imports createEnvironment
                                      from app/models/organization.server
  app/models/organization.server   →  imports getDefaultEnvironmentConcurrencyLimit
                                      from app/services/platform.v3.server

Inside an esbuild __esm bundle, this manifests as:

  init_platform_v3_server() runs init_organization_server() in the
  middle of its body. organization.server's body re-enters
  init_platform_v3_server(), which short-circuits because the outer
  call already cleared its `fn` — so `({ defaultMachine, machines } =
  singleton("machinePresets", ...))` never completes its destructure
  and both vars stay undefined. Object.entries(undefined) crashes
  when `allMachines()` runs inside `createRunEngine()`.

Fix: move the only function in platform.v3.server.ts that imports
from organization.server (`projectCreated`, the sole caller of
`createEnvironment`) into its own file. platform.v3.server.ts no
longer imports from organization.server, so the cycle is gone. Two
trivial supporting changes:

  - export `isCloud` from platform.v3.server (projectCreated needs it)
  - drop the now-unused `Organization` and `Project` type imports

No dynamic imports, no application-code workarounds — just a
structural file split.

Verified:
  - 162/162 e2e.full pass (auth-api, auth-cross-cutting, auth-dashboard)
  - 31/31 api-auth.e2e pass
  - 31/31 @trigger.dev/rbac unit tests pass
  - 7/7 cloud enterprise e2e.full pass against the same webapp build
